Relocate recipe repository
This is a prisma specific repository, hence it should be located under the prisma directory.
