+++ src/application/modules/RecipeModule.ts
@@ -1,7 +1,7 @@
 import { RecipeController } from '@application/api/controllers/RecipeController';
 import { RecipeService } from '@core/services/Recipe';
 import { Module } from '@nestjs/common';
-import { RecipeRepository } from '@infrastructure/adapter/repository/Recipe';
+import { RecipeRepository } from '@infrastructure/adapter/prisma/repository/Recipe';
 import { RecipeTokens } from '@core/domain/di/tokens/Recipe';
 import { PrismaClient } from '@infrastructure/adapter/prisma/client/PrismaClient';
